pointsListPresenter, filtersPresenter separated
