+++ src/presenter/main-presenter.js
@@ -1,11 +1,6 @@
-import { render } from '../framework/render';
-import { generateFilter } from '../mock/mock-filter';
-import FiltersView from '../view/filters-view';
-import PointsListView from '../view/points-list-view';
-import SortView from '../view/sort-view';
-import NoPointsView from '../view/no-points-view';
-import PointPresenter from './point-presenter';
-import { updateItem } from '../utils/common';
+import PointsModel from '../model/points-model';
+import PointsListPresenter from './points-list-presenter';
+import FiltersPresenter from './filters-presenter';
 
 // $======================== MainPresenter ========================$ //
 
@@ -13,87 +8,31 @@ export default class MainPresenter {
   #pointsContainer = null;
   #filtersContainer = null;
 
-  #pointsModel = null;
-  #offersModel = null;
-  #destinationsModel = null;
+  #filtersPresenter = null;
+  #pointsListPresenter = null;
 
+  #pointsModel = new PointsModel();
   #points = [];
 
-  #pointsListComponent = new PointsListView();
-  #sortComponent = new SortView();
-
-  #pointPresenters = new Map();
-
-  constructor({ pointsContainer, filtersContainer, pointsModel, destinationsModel, offersModel }) {
+  constructor({ pointsContainer, filtersContainer }) {
     this.#pointsContainer = pointsContainer;
     this.#filtersContainer = filtersContainer;
-
-    this.#pointsModel = pointsModel;
-    this.#destinationsModel = destinationsModel;
-    this.#offersModel = offersModel;
   }
-
-  #renderPoint(point) {
-    const pointPresenter = new PointPresenter({
-      pointsListComponent: this.#pointsListComponent,
-
-      offersModel: this.#offersModel,
-      destinationsModel: this.#destinationsModel,
-
-      handleDataChange: this.#handlePointChange,
-      handleModeChange: this.#handleModeChange,
-    });
-
-    pointPresenter.init(point);
-    this.#pointPresenters.set(point.id, pointPresenter);
-  }
-
-  #renderPoints() {
-    if (this.#points.length === 0) {
-      this.#renderNoPoints();
-      return;
-    }
-
-    this.#points.forEach((point) => {
-      this.#renderPoint(point);
-    });
-  }
-
-  #renderPointsList() {
-    render(this.#pointsListComponent, this.#pointsContainer);
-    this.#renderPoints();
-  }
-
-  #renderFilters() {
-    const filters = generateFilter(this.#points);
-    render(new FiltersView({ filters }), this.#filtersContainer);
-  }
-
-  #renderSort() {
-    render(this.#sortComponent, this.#pointsContainer);
-  }
-
-  #renderNoPoints() {
-    render(new NoPointsView(), this.#pointsContainer);
-  }
-
-  #handlePointChange = (updatedPoint) => {
-    this.#points = updateItem(this.#points, updatedPoint);
-    this.#pointPresenters.get(updatedPoint.id).init(updatedPoint);
-  };
-
-  #handleModeChange = () => {
-    this.#pointPresenters.forEach((pointPresenter) => {
-      pointPresenter.resetView();
-    });
-  };
 
   init() {
     this.#points = [...this.#pointsModel.points];
 
-    this.#renderFilters();
-    this.#renderSort();
-    this.#renderPointsList();
-  }
+    this.#filtersPresenter = new FiltersPresenter({
+      filtersContainer: this.#filtersContainer,
+      points: this.#points,
+    });
 
+    this.#pointsListPresenter = new PointsListPresenter({
+      pointsContainer: this.#pointsContainer,
+      points: this.#points,
+    });
+
+    this.#filtersPresenter.init();
+    this.#pointsListPresenter.init();
+  }
 }
